Extract Prompt Studio advanced node UI
