+++ web/js/easyuse_anima_prompt_studio.js
@@ -3,7 +3,6 @@ import { easyuseAnimaWatchLocale } from "./easyuse_anima_i18n.js";
 import {
   FIELD_NAMES,
   EXTEND_FIELD_NAMES,
-  ADVANCED_NATIVE_CONTROL_EVENTS,
   DEFAULT_ADVANCED_RESOLUTION_BUCKET,
   DEFAULT_ADVANCED_RESOLUTION_SIZE,
   ADVANCED_WIDGET_INDEX,
@@ -14,14 +13,11 @@ import {
 } from "./prompt_studio/schema.js";
 import {
   getAdvancedEditorElement,
-  setAdvancedEditorElement,
-  setAdvancedFields,
 } from "./prompt_studio/state.js";
 import {
   stopAdvancedControlEvent,
 } from "./prompt_studio/dom.js";
 import {
-  forwardAdvancedWheelToCanvas,
   installMiddlePanForwarder,
 } from "./prompt_studio/canvas_forwarding.js";
 import {
@@ -46,9 +42,6 @@ import {
   syncAdvancedNodes,
 } from "./prompt_studio/node_hooks.js";
 import {
-  ensureAdvancedStyle,
-} from "./prompt_studio/style.js";
-import {
   applyPromptStudioSettings,
   applyPromptStudioTextStyle,
   loadPromptStudioSettings,
@@ -57,19 +50,8 @@ import {
   hideTrainedTagTooltip,
 } from "./prompt_studio/tooltip.js";
 import {
-  createAdvancedControlBar,
-  createAdvancedResolutionBar,
-  createAdvancedWildcardBar,
-} from "./prompt_studio/advanced_controls.js";
-import {
-  createAdvancedPane,
-} from "./prompt_studio/advanced_fields_ui.js";
-import {
-  advancedFieldLabel,
   advancedWidget,
-  applyAdvancedNaiaGeneralAutoToggle,
   hideAdvancedControlWidgets as hideAdvancedControlWidgetsWithHooks,
-  hideAdvancedInternalWidget,
   parseAdvancedFields,
   removeAdvancedInternalInputSockets,
   repairAdvancedInternalWidgetValues as repairAdvancedInternalWidgetValuesWithHooks,
@@ -80,10 +62,7 @@ import {
   refreshAllPromptHighlights,
 } from "./prompt_studio/highlight.js";
 import {
-  registerAdvancedAutocompleteInput,
-  scheduleAdvancedFieldHighlight,
   scheduleAdvancedHighlights,
-  updateAdvancedFieldHighlight,
 } from "./prompt_studio/advanced_highlights.js";
 import {
   updateHighlight,
@@ -92,17 +71,12 @@ import {
   findWidget,
 } from "./prompt_studio/widgets.js";
 import {
-  advancedEditorMinimumHeight,
-  advancedEditorWidgetHeight,
   updateAdvancedEditorWidth,
 } from "./prompt_studio/layout.js";
 import {
   scheduleAdvancedLayout as scheduleAdvancedLayoutWithHooks,
   scheduleAdvancedResizeFinalize as scheduleAdvancedResizeFinalizeWithHooks,
 } from "./prompt_studio/advanced_layout_controller.js";
-import {
-  guardAdvancedEditorNativeControlEvent,
-} from "./prompt_studio/wheel.js";
 import {
   expandStudioInputToContent as expandStudioInputToContentWithHooks,
   growStudioManualHeightToContent as growStudioManualHeightToContentWithHooks,
@@ -124,6 +98,11 @@ import {
   hookStudioNode as hookStudioNodeWithHooks,
 } from "./prompt_studio/studio_node_ui.js";
 import {
+  hookAdvancedNode as hookAdvancedNodeWithHooks,
+  renderAdvancedEditor as renderAdvancedEditorWithHooks,
+  scheduleHookAdvancedNode as scheduleHookAdvancedNodeWithHooks,
+} from "./prompt_studio/advanced_node_ui.js";
+import {
   applyAdvancedExecutedInputs as applyAdvancedExecutedInputsWithHooks,
   syncAdvancedValues as syncAdvancedValuesWithHooks,
 } from "./prompt_studio/advanced_values.js";
@@ -132,7 +111,6 @@ import {
 } from "./prompt_studio/wildcard_values.js";
 import {
   captureAdvancedConfigure,
-  ensureAdvancedWidgetValue,
   pruneDisconnectedAdvancedFieldInputValues,
 } from "./prompt_studio/serialization.js";
 
@@ -285,27 +263,6 @@ function promptHighlightHooks() {
   };
 }
 
-function advancedControlHooks() {
-  return {
-    renderAdvancedEditor,
-    scheduleAdvancedHighlights,
-    scheduleAdvancedLayout,
-  };
-}
-
-function advancedFieldsUiHooks() {
-  return {
-    advancedFieldLabel,
-    applyAdvancedNaiaGeneralAutoToggle,
-    parseAdvancedFields,
-    registerAdvancedAutocompleteInput,
-    scheduleAdvancedFieldHighlight,
-    scheduleAdvancedLayout,
-    updateAdvancedFieldHighlight,
-    writeAdvancedFields,
-  };
-}
-
 function markGraphDirty() {
   app.graph?.setDirtyCanvas?.(true, true);
 }
@@ -369,82 +326,30 @@ function hookStudioNode(node, attempt = 0) {
   });
 }
 
+function installAdvancedSaveSyncForApp() {
+  installAdvancedSaveSync(app, syncAllAdvancedNodes);
+}
+
+function advancedNodeUiHooks() {
+  return {
+    hideAdvancedControlWidgets,
+    installAdvancedSaveSync: installAdvancedSaveSyncForApp,
+    scheduleAdvancedHighlights,
+    scheduleAdvancedLayout,
+    writeAdvancedFields,
+  };
+}
+
 function renderAdvancedEditor(node) {
-  const editor = getAdvancedEditorElement(node);
-  if (!editor) {
-    return;
-  }
-  const fields = setAdvancedFields(node, parseAdvancedFields(node));
-  applyAdvancedNaiaGeneralAutoToggle(node, fields);
-  editor.innerHTML = "";
-  updateAdvancedEditorWidth(node);
-  const panes = document.createElement("div");
-  panes.className = "easyuse-anima-advanced-panes";
-  const fieldHooks = advancedFieldsUiHooks();
-  panes.append(
-    createAdvancedPane(node, "positive", "advanced.positivePrompt", fieldHooks),
-    createAdvancedPane(node, "negative", "advanced.negativePrompt", fieldHooks),
-  );
-  editor.append(
-    createAdvancedControlBar(node, advancedControlHooks()),
-    createAdvancedWildcardBar(node, advancedControlHooks()),
-    createAdvancedResolutionBar(node, advancedControlHooks()),
-    panes,
-  );
-  writeAdvancedFields(node, fields);
-  scheduleAdvancedLayout(node, "render");
+  renderAdvancedEditorWithHooks(node, advancedNodeUiHooks());
 }
 
 function hookAdvancedNode(node) {
-  ensureAdvancedStyle();
-  installAdvancedSaveSync(app, syncAllAdvancedNodes);
-  ensureAdvancedWidgetValue(node, advancedWidget(node));
-  removeAdvancedInternalInputSockets(node);
-  hideAdvancedInternalWidget(node, "advanced_fields");
-  hideAdvancedControlWidgets(node);
-  node.serialize_widgets = true;
-  node.minWidth = Math.max(Number(node.minWidth) || 0, 360);
-  if (Array.isArray(node.size)) {
-    node.size[0] = Math.max(Number(node.size[0]) || 420, 360);
-  }
-  if (!getAdvancedEditorElement(node)) {
-    const editor = document.createElement("div");
-    editor.className = "easyuse-anima-advanced-editor";
-    editor.addEventListener("wheel", forwardAdvancedWheelToCanvas, { capture: true, passive: false });
-    for (const eventName of ADVANCED_NATIVE_CONTROL_EVENTS) {
-      editor.addEventListener(eventName, guardAdvancedEditorNativeControlEvent);
-    }
-    setAdvancedEditorElement(node, editor);
-    const widget = node.addDOMWidget?.("easyuse_anima_advanced_editor", "EasyUseAnimaAdvancedEditor", editor, {
-      serialize: false,
-      hideOnZoom: false,
-      getMinHeight: () => advancedEditorMinimumHeight(node),
-      getHeight: () => advancedEditorWidgetHeight(node),
-    });
-    if (widget) {
-      node.__easyuseAnimaAdvancedDomWidget = widget;
-      widget.computeLayoutSize = () => {
-        const height = advancedEditorWidgetHeight(node);
-        return {
-          minHeight: advancedEditorMinimumHeight(node),
-          height,
-          minWidth: 280,
-        };
-      };
-    }
-  }
-  renderAdvancedEditor(node);
+  hookAdvancedNodeWithHooks(node, advancedNodeUiHooks());
 }
 
 function scheduleHookAdvancedNode(node) {
-  if (!node || node.__easyuseAnimaAdvancedHookScheduled) {
-    return;
-  }
-  node.__easyuseAnimaAdvancedHookScheduled = true;
-  requestAnimationFrame(() => {
-    node.__easyuseAnimaAdvancedHookScheduled = false;
-    hookAdvancedNode(node);
-  });
+  scheduleHookAdvancedNodeWithHooks(node, advancedNodeUiHooks());
 }
 
 const syncAllAdvancedNodes = () => syncAdvancedNodes(app, syncAdvancedValues);
